# docs/swagger.yaml
definitions:
...
  entities.Organization:
    properties:
      api_key:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: integer
      type:
        type: string
    type: object
  entities.Permission:
    properties:
      action:
        type: string
      description:
        type: string
      endpoint:
        type: string
      id:
        type: string
      name:
        type: string
      subject:
        type: string
    type: object
...
  entities.Role:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/entities.Organization'
      organization_id:
        type: string
      permissions:
        items:
          $ref: '#/definitions/entities.Permission'
        type: array
      type:
        type: string
    type: object
...
  requests.AccountRoleRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
...
  requests.RolePermissionsRequest:
    properties:
      permissions:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - permissions
    type: object
...
  requests.UpsertPermissionRequest:
    properties:
      action:
        type: string
      description:
        type: string
      endpoint:
        type: string
      name:
        type: string
      subject:
        type: string
    required:
    - action
    - description
    - endpoint
    - name
    - subject
    type: object
  requests.UpsertRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        type: string
    required:
    - description
    - name
    type: object
...
  response.Response:
    properties:
      data: {}
      meta: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/accounts/{id}/roles:
    delete:
      consumes:
      - application/json
      description: Revoke role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: account_id
        in: path
        name: id
        required: true
        type: string
      - description: AccountRoleRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.AccountRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Revoke role.
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Add role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: account_id
        in: path
        name: id
        required: true
        type: string
      - description: AccountRoleRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.AccountRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Add role.
      tags:
      - accounts
...
  /api/permissions:
    get:
      consumes:
      - application/json
      description: Dashboard.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: rows limit
        in: query
        name: limit
        required: true
        type: integer
      - description: rows offset
        in: query
        name: offset
        required: true
        type: integer
      - description: order field
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Permission'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get permissions list.
      tags:
      - permission
    post:
      consumes:
      - application/json
      description: Create permission.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreatePermissionRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpsertPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Permission'
              type: object
      security:
      - X-Authenticate: []
      summary: Create new permission.
      tags:
      - permission
  /api/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete permission.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: permission_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete permission.
      tags:
      - permission
    get:
      consumes:
      - application/json
      description: Get permission.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: permission_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Permission'
              type: object
      security:
      - X-Authenticate: []
      summary: Get permission.
      tags:
      - permission
    put:
      consumes:
      - application/json
      description: Update permission.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: permission_id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatePermissionRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpsertPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Permission'
              type: object
      security:
      - X-Authenticate: []
      summary: Update new permission.
      tags:
      - permission
...
  /api/roles:
    get:
      consumes:
      - application/json
      description: Available backoffice roles.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: rows limit
        in: query
        name: limit
        required: true
        type: integer
      - description: rows offset
        in: query
        name: offset
        required: true
        type: integer
      - description: order field
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Role'
                  type: array
              type: object
      security:
      - X-Authenticate: []
      summary: Get roles list.
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: CreateRoleRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpsertRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Role'
              type: object
      security:
      - X-Authenticate: []
      summary: Add new role.
      tags:
      - roles
  /api/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: role_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - X-Authenticate: []
      summary: Delete role.
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get existing role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: role_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Role'
              type: object
      security:
      - X-Authenticate: []
      summary: Get role.
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update existing role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: role_id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateRoleRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.UpsertRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/entities.Role'
              type: object
      security:
      - X-Authenticate: []
      summary: Update role.
      tags:
      - roles
  /api/roles/{id}/permissions:
    delete:
      consumes:
      - application/json
      description: Delete permissions from role.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: role_id
        in: path
        name: id
        required: true
        type: string
      - description: RolePermissionsRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.RolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Revoke permissions.
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Add permissions.
      parameters:
      - description: Insert your access token
        in: header
        name: X-Auth
        required: true
        type: string
      - description: role_id
        in: path
        name: id
        required: true
        type: string
      - description: RolePermissionsRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.RolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - X-Authenticate: []
      summary: Add permissions.
      tags:
      - roles
...
swagger: "2.0"